requirements.txt
discord.py
psycopg2
praw
itertools